AIComplaintBackend/AiApp/requirements.txt
Django==5.2.3
pymongo==4.6.1
python-dotenv==1.1.0
requests==2.31.0
groq==0.5.0
